+++ prototype/mobile_endpoint/views/receiver.py
@@ -1,7 +1,8 @@
 from flask import request
+from mobile_endpoint.backends.sql.dao import SQLDao
 
 from mobile_endpoint.case.case_processing import process_cases_in_form
-from mobile_endpoint.dao import SQLDao
+
 from mobile_endpoint.extensions import requires_auth
 from mobile_endpoint.form.form_processing import create_xform, get_instance_and_attachments, get_request_metadata
 from mobile_endpoint.views import ota_mod
