+++ prototype/tests/mock.py
@@ -5,8 +5,8 @@ from datetime import datetime
 from uuid import uuid4
 from xml.etree import ElementTree
 from flask.templating import render_template_string
+from mobile_endpoint.backends.sql.dao import SQLDao
 from mobile_endpoint.case.xml import NS_VERSION_MAP, V2
-from mobile_endpoint.dao import SQLDao
 from mobile_endpoint.utils import json_format_datetime
